+++ webapp/src/main/java/ee/cyber/cdoc2/Cdoc2RpServerApplication.java
@@ -5,6 +5,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
+import ee.cyber.cdoc2.server.config.MonitoringUtil;
+
 @SpringBootApplication()
 @ConfigurationPropertiesScan
 @EnableScheduling
@@ -14,6 +16,9 @@ public final class Cdoc2RpServerApplication {
     }
 
     public static void main(String[] args) {
-        SpringApplication.run(Cdoc2RpServerApplication.class, args);
+        SpringApplication app = new SpringApplication(Cdoc2RpServerApplication.class);
+        // capture startup events for startup actuator endpoint
+        app.setApplicationStartup(MonitoringUtil.getApplicationStartupInfo());
+        app.run(args);
     }
 }