+++ webapp/src/main/java/ee/cyber/cdoc2/server/config/SecurityConfiguration.java
@@ -1,0 +1,71 @@
+package ee.cyber.cdoc2.server.config;
+
+import lombok.RequiredArgsConstructor;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.Customizer;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.provisioning.InMemoryUserDetailsManager;
+import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.authentication.preauth.x509.SubjectX500PrincipalExtractor;
+
+
+/**
+ * Security configuration
+ */
+@Configuration
+@EnableWebSecurity
+@RequiredArgsConstructor
+@SuppressWarnings("java:S4502")
+public class SecurityConfiguration {
+
+    private final MonitoringConfigProperties credentials;
+
+    @Bean
+    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+        http
+            .csrf(AbstractHttpConfigurer::disable)
+            .authorizeHttpRequests(authorize ->
+                authorize
+                    .requestMatchers("/session_nonce").permitAll()
+                    .requestMatchers("/.well-known/**").permitAll()
+                    .requestMatchers("/mid/**").permitAll()
+                    .requestMatchers("/sid/**").permitAll()
+                    .requestMatchers("/info").permitAll()
+                    // authenticated URI must go first
+                    .requestMatchers("/actuator/prometheus").authenticated()
+                    .requestMatchers("/actuator/**").permitAll()
+                    .anyRequest().authenticated()
+            )
+            .x509(x509 ->
+                x509
+                    // extracts the CN attribute from the certificate's subject DN
+                    .x509PrincipalExtractor(new SubjectX500PrincipalExtractor())
+                    .userDetailsService(userDetailsService())
+            )
+            .sessionManagement(sessionManagementConfigurer ->
+                sessionManagementConfigurer.sessionCreationPolicy(SessionCreationPolicy.NEVER)
+            )
+            .httpBasic(Customizer.withDefaults());
+
+        return http.build();
+    }
+
+    @Bean
+    public UserDetailsService userDetailsService() {
+        UserDetails user = User.builder()
+            .username(credentials.username())
+            .password("{noop}" + credentials.password())
+            .roles("USER")
+            .build();
+        return new InMemoryUserDetailsManager(user);
+    }
+
+}